Ignore `window.Vue` in ES2015 Environments
I'm not sure if this will apply to everyone; in fact, it could possibly break many installations. However, it seems to reason that if `isES2015 == true`, since assets are being compiled in webpack or rollup, there is no reason for this library to search for a global instance of Vue.

This fixes #12.